Issue #2335 Test Case
In this test case we import 3 routes with /foo prefix, and then create
a 4th route sharing the same prefix. The result is the 4th route being
dropped by the compiler and a "Route "testroute_4" does not exist."
exception thrown. Modifying "testroute_4" to have a pattern like
"/afoo/bar4" makes it work, proving the bug.